pyproject.toml
[tool.poetry]
name = "tomato"
version = "0.0.1-alpha"
description = "Plaintext-type ledger tool"
authors = ["Warner Leijenaar <git@renraw.nl>"]
readme = "README.md"
packages = [
  {include = "tomato", from = "src"},
  {include = "tomato/py.typed", from = "src"},
]
include = [
  {path = "var/etc/defaults.toml"}
]


[tool.poetry.urls]
"Homepage" = "https://github.com/renraw-nl/tomato"


[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.7.0"
shellingham = "^1.5.0.post1"
python-dotenv = "^0.21.0"
tomlkit = "^0.11.6"
structlog = "^22.3.0"
python-json-logger = "^2.0.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.12.0"
isort = "^5.11.4"
ruff = "^0.0.213"
mypy = "^0.991"
pre-commit = "^2.21.0"


[build-system]
requires = ["poetry-core>=1.3.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[tool.poetry.scripts]
tomato = 'tomato.api.cli:app'
# gui = 'office_unlocker.gui:run'

[tool.pytest.ini_options]
# asyncio_mode = "auto"
# addopts = "--no-cov-on-fail --showlocals --strict-markers"
addopts = "--showlocals --strict-markers"
testpaths = ["tests"]

[tool.mypy]
allow_redefinition = true
disallow_any_generics = false
disallow_subclassing_any = true
disallow_untyped_calls = false
disallow_untyped_defs = true
implicit_reexport = true
no_implicit_optional = true
show_error_codes = true
strict = true
strict_equality = true
strict_optional = false
warn_redundant_casts = true
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = true

[tool.black]
line-length = 88
target-version = ["py311"]
exclude = '''
/(
   \.git
  | \.ruff_cache
  | \.mypy_cache
  | \.vscode
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = "builder, tests"
line_length = 88
multi_line_output = 3
no_lines_before = "LOCALFOLDER"
order_by_type = false
reverse_relative = true
profile = "black"
